replace distributed_model_parallel_state with dist_environment
