+++ sub-packages/bionemo-testing/tests/bionemo/testing/test_megatron_parallel_state_utils.py
@@ -13,12 +13,97 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pytest
 import torch
 import torch.distributed as dist
 from megatron.core import parallel_state
 from nemo import lightning as nl
 
 from bionemo.testing import megatron_parallel_state_utils
+
+
+MAX_WORLD_SIZE = 4
+AVAILABLE_WORLD_SIZE = torch.cuda.device_count()
+WORLD_SIZES = range(min(MAX_WORLD_SIZE, torch.cuda.device_count()))
+
+
+def _test_all_reduce_sum(rank: int, world_size: int):
+    """Private test function for torch.distributed mean reduce."""
+    with megatron_parallel_state_utils.distributed_model_parallel_state(rank=rank, world_size=world_size):
+        tensor = torch.tensor([rank + 1]).cuda(rank)
+        dist.all_reduce(tensor)
+        assert tensor.item() == world_size * (world_size + 1) / 2
+
+
+@pytest.mark.parametrize("world_size", WORLD_SIZES)
+def test_all_reduce_sum(world_size: int):
+    """Multiprocessing test of _test_all_reduce_sum."""
+    torch.multiprocessing.spawn(
+        fn=_test_all_reduce_sum,
+        args=(world_size,),
+        nprocs=world_size,
+    )
+
+
+def _test_data_parallel_group(rank: int, world_size: int):
+    """Private test function for dp parallel state."""
+    with megatron_parallel_state_utils.distributed_model_parallel_state(rank=rank, world_size=world_size):
+        assert parallel_state.get_data_parallel_rank() == rank
+        assert parallel_state.get_data_parallel_world_size() == world_size
+        assert parallel_state.get_data_parallel_src_rank() == 0
+
+
+@pytest.mark.parametrize("world_size", WORLD_SIZES)
+def test_data_parallel_group(world_size: int):
+    """Multiprocessing test of _test_data_parallel_group."""
+    torch.multiprocessing.spawn(
+        fn=_test_data_parallel_group,
+        args=(world_size,),
+        nprocs=world_size,
+    )
+
+
+def _test_tensor_model_parallel_group(rank: int, world_size: int):
+    """Private test function for tp parallel state."""
+    with megatron_parallel_state_utils.distributed_model_parallel_state(
+        rank=rank, world_size=world_size, tensor_model_parallel_size=world_size
+    ):
+        assert parallel_state.get_tensor_model_parallel_rank() == rank
+        assert parallel_state.get_tensor_model_parallel_world_size() == world_size
+        assert parallel_state.get_tensor_model_parallel_src_rank() == 0
+
+
+@pytest.mark.parametrize("world_size", WORLD_SIZES)
+def test_tensor_model_parallel_group(world_size: int):
+    """Multiprocessing test of _test_tensor_model_parallel_group."""
+    torch.multiprocessing.spawn(
+        fn=_test_tensor_model_parallel_group,
+        args=(world_size,),
+        nprocs=world_size,
+    )
+
+
+def _test_pipeline_model_parallel_group(rank: int, world_size: int):
+    """Private test function for pp parallel state."""
+    with megatron_parallel_state_utils.distributed_model_parallel_state(
+        rank=rank, world_size=world_size, pipeline_model_parallel_size=world_size
+    ):
+        assert parallel_state.get_pipeline_model_parallel_rank() == rank
+        assert parallel_state.get_pipeline_model_parallel_world_size() == world_size
+        if rank == 0:
+            assert parallel_state.is_pipeline_first_stage()
+        if rank == world_size:
+            assert parallel_state.is_pipeline_last_stage()
+
+
+@pytest.mark.parametrize("world_size", WORLD_SIZES)
+def test_pipeline_model_parallel_group(world_size: int):
+    """Multiprocessing test of _test_pipeline_model_parallel_group."""
+    torch.multiprocessing.spawn(
+        fn=_test_pipeline_model_parallel_group,
+        args=(world_size,),
+        nprocs=world_size,
+    )
 
 
 def test_load_megatron_strategy():
@@ -40,7 +125,7 @@ def test_construct_nemo_lightning_trainer():
         assert trainer.max_steps == 5
 
 
-def test_rank0_first_pipeline():
+def test_mock_rank0_first_pipeline():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(
         world_size=8, rank=0, pipeline_model_parallel_size=8
     ):
@@ -48,7 +133,7 @@ def test_rank0_first_pipeline():
         assert not parallel_state.is_pipeline_last_stage()
 
 
-def test_rank4_mid_pipeline():
+def test_mock_rank4_mid_pipeline():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(
         world_size=8, rank=4, pipeline_model_parallel_size=8
     ):
@@ -56,7 +141,7 @@ def test_rank4_mid_pipeline():
         assert not parallel_state.is_pipeline_last_stage()
 
 
-def test_rank7_last_pipeline():
+def test_mock_rank7_last_pipeline():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(
         world_size=8, rank=7, pipeline_model_parallel_size=8
     ):
@@ -64,22 +149,22 @@ def test_rank7_last_pipeline():
         assert parallel_state.is_pipeline_last_stage()
 
 
-def test_get_pp_group():
+def test_mock_get_pp_group():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, pipeline_model_parallel_size=2):
         assert parallel_state.get_pipeline_model_parallel_group() is not None
 
 
-def test_get_tp_group():
+def test_mock_get_tp_group():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, tensor_model_parallel_size=2):
         assert parallel_state.get_tensor_model_parallel_group() is not None
 
 
-def test_get_cp_group():
+def test_mock_get_cp_group():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, context_parallel_size=2):
         assert parallel_state.get_context_parallel_group() is not None
 
 
-def test_all_reduce():
+def test_mock_all_reduce():
     # Adapted from https://github.com/pytorch/pytorch/blob/main/test/distributed/test_fake_pg.py
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, rank=1):
         output = torch.ones(3, 3).cuda() * dist.get_rank()
@@ -87,7 +172,7 @@ def test_all_reduce():
         assert tuple(output.shape) == (3, 3)
 
 
-def test_allgather():
+def test_mock_allgather():
     # Adapted from https://github.com/pytorch/pytorch/blob/main/test/distributed/test_fake_pg.py
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, rank=1):
         input_tensor = torch.ones(3, 3) * dist.get_rank()
@@ -97,7 +182,7 @@ def test_allgather():
             assert tuple(out_tensor.shape) == (3, 3)
 
 
-def test_reduce_scatter():
+def test_mock_reduce_scatter():
     # Adapted from https://github.com/pytorch/pytorch/blob/main/test/distributed/test_fake_pg.py
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, rank=1):
         to_reduce_scatter = [torch.ones(3, 3) * rank for rank in range(2)]
@@ -107,7 +192,7 @@ def test_reduce_scatter():
         assert tuple(output_tensor.shape) == (3, 3)
 
 
-def test_all_reduce_sum():
+def test_mock_all_reduce_sum():
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, rank=1):
         tensor = torch.tensor([dist.get_rank() + 1])
         dist.all_reduce(tensor)
